Stand 29.09.2022

+++ MyUserMonitoringSettings.user.cs
@@ -10,7 +10,7 @@ namespace MyUserMonitoringSettings
     [Serializable]
     public struct MY_USER_MONITORING_DATA_T
     {
-        public UInt64[] UserMonitoringTimeBase_ms { get; set; }     // Signal name
+        public int[] UserMonitoringTimeBase_ms { get; set; }     // Signal name
         public bool[,] UsedInChart { get; set; }     // Flag, that signal/trigger is used in a chart
         public int [,] SeriesInChart { get; set; }     // signals/triggers is used in a chart
         public Color ChartBColor { get; set; }      // Chart background color
@@ -24,7 +24,7 @@ namespace MyUserMonitoringSettings
     {
         public void SetDefault()
         {
-            settings.UserMonitoringTimeBase_ms = new UInt64[2];
+            settings.UserMonitoringTimeBase_ms = new int[2];
             settings.UsedInChart = new bool[2, RCmanager.Constants.MODULES * RCmanager.Constants.CHANNELS + RCmanager.Constants.IRQ_IOS];
             settings.SeriesInChart = new int[2, RCmanager.Constants.MODULES * RCmanager.Constants.CHANNELS + RCmanager.Constants.IRQ_IOS];
         }
